refactor(supervisor): rename runtime.ts -> supervisorRuntime.ts
Breaks the file-vs-directory naming collision: src/supervisor/runtime.ts sat
beside src/supervisor/runtime/, so "@/supervisor/runtime" was ambiguous to grep,
go-to-definition, and newcomers. The SupervisorRuntime god-facade is now
supervisorRuntime.ts; "@/supervisor/runtime" unambiguously means the directory.
Updates 6 importers. The wsl/runtime module is unrelated and untouched. Pure
rename; typecheck green. The full dissolve of the delegate facade (~110 one-line
delegates) is deferred (requires reworking 31 integration tests + the typed
SupervisorIpcHandlerMap).
